Allow for chaining multible animations
This allows for chaining multible animations by calling the chain(key) method multible times like this:
this.mole.anims.play('digging',true).anims.chain('lifting').anims.chain('looking').anims.chain('lowering');

This commit adds a new property nextAnimsQueue. This is an array that holds the chained keys in order of chain() requests.
This list will be worked on in the stop() function.